Refactor: 수익률의 변수 타입 변경

+++ src/main/java/lotto/view/OutputView.java
@@ -58,7 +58,7 @@ public class OutputView {
         System.out.println(result);
     }
 
-    public static void printRateOfProfit(long rateOfProfit) {
+    public static void printRateOfProfit(double rateOfProfit) {
 
         StringBuilder result = new StringBuilder();
